+++ casparser/__init__.py
@@ -9,4 +9,4 @@ __all__ = [
     "CapitalGainsReport",
 ]
 
-__version__ = "1.0.0"
+__version__ = "1.0.1"